move gpool to network folder
